CachedInterpreter: Move to its own directory
